+++ Javadoc/constant-values.html
@@ -89,28 +89,28 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a name="deltadak.Controller.MAX_COLUMNS">
+<td class="colFirst"><a name="deltadak.ui.Controller.MAX_COLUMNS">
 <!--   -->
 </a><code>private&nbsp;static&nbsp;final&nbsp;int</code></td>
 <td><code><a href="deltadak/Controller.html#MAX_COLUMNS">MAX_COLUMNS</a></code></td>
 <td class="colLast"><code>3</code></td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a name="deltadak.Controller.MAX_LIST_LENGTH">
+<td class="colFirst"><a name="deltadak.ui.Controller.MAX_LIST_LENGTH">
 <!--   -->
 </a><code>private&nbsp;static&nbsp;final&nbsp;int</code></td>
 <td><code><a href="deltadak/Controller.html#MAX_LIST_LENGTH">MAX_LIST_LENGTH</a></code></td>
 <td class="colLast"><code>7</code></td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a name="deltadak.Controller.NUMBER_OF_DAYS">
+<td class="colFirst"><a name="deltadak.ui.Controller.NUMBER_OF_DAYS">
 <!--   -->
 </a><code>private&nbsp;static&nbsp;final&nbsp;int</code></td>
 <td><code><a href="deltadak/Controller.html#NUMBER_OF_DAYS">NUMBER_OF_DAYS</a></code></td>
 <td class="colLast"><code>9</code></td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a name="deltadak.Controller.NUMBER_OF_MOVING_DAYS">
+<td class="colFirst"><a name="deltadak.ui.Controller.NUMBER_OF_MOVING_DAYS">
 <!--   -->
 </a><code>private&nbsp;static&nbsp;final&nbsp;int</code></td>
 <td><code><a href="deltadak/Controller.html#NUMBER_OF_MOVING_DAYS">NUMBER_OF_MOVING_DAYS</a></code></td>
