+++ Javadoc/deltadak/Controller.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>deltadak.Controller</li>
+<li>deltadak.ui.Controller</li>
 </ul>
 </li>
 </ul>
@@ -237,7 +237,7 @@ implements javafx.fxml.Initializable</pre>
 </tr>
 <tr id="i6" class="altColor">
 <td class="colFirst"><code>(package private) void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../deltadak/Controller.html#createContextMenu-javafx.scene.input.MouseEvent-deltadak.LabelCell-javafx.scene.control.ListView-java.time.LocalDate-">createContextMenu</a></span>(javafx.scene.input.MouseEvent&nbsp;event,
+<td class="colLast"><code><span class="memberNameLink"><a href="../deltadak/Controller.html#createContextMenu-javafx.scene.input.MouseEvent-deltadak.ui.LabelCell-javafx.scene.control.ListView-java.time.LocalDate-">createContextMenu</a></span>(javafx.scene.input.MouseEvent&nbsp;event,
                  <a href="../deltadak/LabelCell.html" title="class in deltadak">LabelCell</a>&nbsp;labelCell,
                  javafx.scene.control.ListView&lt;<a href="../deltadak/Task.html" title="class in deltadak">Task</a>&gt;&nbsp;listViewItems,
                  java.time.LocalDate&nbsp;day)</code>
@@ -302,7 +302,7 @@ implements javafx.fxml.Initializable</pre>
 </tr>
 <tr id="i16" class="altColor">
 <td class="colFirst"><code>private javafx.scene.control.Menu</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../deltadak/Controller.html#makeRepeatMenu-deltadak.LabelCell-java.time.LocalDate-">makeRepeatMenu</a></span>(<a href="../deltadak/LabelCell.html" title="class in deltadak">LabelCell</a>&nbsp;labelCell,
+<td class="colLast"><code><span class="memberNameLink"><a href="../deltadak/Controller.html#makeRepeatMenu-deltadak.ui.LabelCell-java.time.LocalDate-">makeRepeatMenu</a></span>(<a href="../deltadak/LabelCell.html" title="class in deltadak">LabelCell</a>&nbsp;labelCell,
               java.time.LocalDate&nbsp;day)</code>
 <div class="block">Makes the menu with options to repeat for 1-8 weeks.</div>
 </td>
@@ -323,7 +323,7 @@ implements javafx.fxml.Initializable</pre>
 </tr>
 <tr id="i19" class="rowColor">
 <td class="colFirst"><code>private void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../deltadak/Controller.html#setBackgroundColor-javafx.scene.control.MenuItem-deltadak.LabelCell-">setBackgroundColor</a></span>(javafx.scene.control.MenuItem&nbsp;menuItem,
+<td class="colLast"><code><span class="memberNameLink"><a href="../deltadak/Controller.html#setBackgroundColor-javafx.scene.control.MenuItem-deltadak.ui.LabelCell-">setBackgroundColor</a></span>(javafx.scene.control.MenuItem&nbsp;menuItem,
                   <a href="../deltadak/LabelCell.html" title="class in deltadak">LabelCell</a>&nbsp;labelCell)</code>
 <div class="block">sets the background color of a LabelCell</div>
 </td>
@@ -417,7 +417,7 @@ implements javafx.fxml.Initializable</pre>
 <pre>private static final&nbsp;int NUMBER_OF_DAYS</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../constant-values.html#deltadak.Controller.NUMBER_OF_DAYS">Constant Field Values</a></dd>
+<dd><a href="../constant-values.html#deltadak.ui.Controller.NUMBER_OF_DAYS">Constant Field Values</a></dd>
 </dl>
 </li>
 </ul>
@@ -430,7 +430,7 @@ implements javafx.fxml.Initializable</pre>
 <pre>private static final&nbsp;int MAX_COLUMNS</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../constant-values.html#deltadak.Controller.MAX_COLUMNS">Constant Field Values</a></dd>
+<dd><a href="../constant-values.html#deltadak.ui.Controller.MAX_COLUMNS">Constant Field Values</a></dd>
 </dl>
 </li>
 </ul>
@@ -443,7 +443,7 @@ implements javafx.fxml.Initializable</pre>
 <pre>private static final&nbsp;int MAX_LIST_LENGTH</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../constant-values.html#deltadak.Controller.MAX_LIST_LENGTH">Constant Field Values</a></dd>
+<dd><a href="../constant-values.html#deltadak.ui.Controller.MAX_LIST_LENGTH">Constant Field Values</a></dd>
 </dl>
 </li>
 </ul>
@@ -474,7 +474,7 @@ implements javafx.fxml.Initializable</pre>
 <pre>private static final&nbsp;int NUMBER_OF_MOVING_DAYS</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../constant-values.html#deltadak.Controller.NUMBER_OF_MOVING_DAYS">Constant Field Values</a></dd>
+<dd><a href="../constant-values.html#deltadak.ui.Controller.NUMBER_OF_MOVING_DAYS">Constant Field Values</a></dd>
 </dl>
 </li>
 </ul>
@@ -696,7 +696,7 @@ implements javafx.fxml.Initializable</pre>
 <div class="block">refreshes all listviews using data from the database</div>
 </li>
 </ul>
-<a name="makeRepeatMenu-deltadak.LabelCell-java.time.LocalDate-">
+<a name="makeRepeatMenu-deltadak.ui.LabelCell-java.time.LocalDate-">
 <!--   -->
 </a>
 <ul class="blockList">
@@ -714,7 +714,7 @@ implements javafx.fxml.Initializable</pre>
 </dl>
 </li>
 </ul>
-<a name="createContextMenu-javafx.scene.input.MouseEvent-deltadak.LabelCell-javafx.scene.control.ListView-java.time.LocalDate-">
+<a name="createContextMenu-javafx.scene.input.MouseEvent-deltadak.ui.LabelCell-javafx.scene.control.ListView-java.time.LocalDate-">
 <!--   -->
 </a>
 <ul class="blockList">
@@ -734,7 +734,7 @@ implements javafx.fxml.Initializable</pre>
 </dl>
 </li>
 </ul>
-<a name="setBackgroundColor-javafx.scene.control.MenuItem-deltadak.LabelCell-">
+<a name="setBackgroundColor-javafx.scene.control.MenuItem-deltadak.ui.LabelCell-">
 <!--   -->
 </a>
 <ul class="blockList">
