+++ Javadoc/deltadak/LabelCell.html
@@ -125,7 +125,7 @@ var activeTableTab = "activeTableTab";
 <li>javafx.scene.control.cell.TextFieldListCell&lt;<a href="../deltadak/Task.html" title="class in deltadak">Task</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>deltadak.LabelCell</li>
+<li>deltadak.ui.LabelCell</li>
 </ul>
 </li>
 </ul>
@@ -301,7 +301,7 @@ extends javafx.scene.control.cell.TextFieldListCell&lt;<a href="../deltadak/Task
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../deltadak/LabelCell.html#LabelCell-deltadak.Controller-">LabelCell</a></span>(<a href="../deltadak/Controller.html" title="class in deltadak">Controller</a>&nbsp;controller)</code>
+<td class="colOne"><code><span class="memberNameLink"><a href="../deltadak/LabelCell.html#LabelCell-deltadak.ui.Controller-">LabelCell</a></span>(<a href="../deltadak/Controller.html" title="class in deltadak">Controller</a>&nbsp;controller)</code>
 <div class="block">Constructor.</div>
 </td>
 </tr>
@@ -537,7 +537,7 @@ extends javafx.scene.control.cell.TextFieldListCell&lt;<a href="../deltadak/Task
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="LabelCell-deltadak.Controller-">
+<a name="LabelCell-deltadak.ui.Controller-">
 <!--   -->
 </a>
 <ul class="blockListLast">
