+++ Javadoc/deltadak/TaskConverter.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li>javafx.util.StringConverter&lt;<a href="../deltadak/Task.html" title="class in deltadak">Task</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>deltadak.TaskConverter</li>
+<li>deltadak.ui.TaskConverter</li>
 </ul>
 </li>
 </ul>
