+++ Javadoc/index-files/index-14.html
@@ -72,7 +72,7 @@
 </a>
 <h2 class="title">S</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../deltadak/Controller.html#setBackgroundColor-javafx.scene.control.MenuItem-deltadak.LabelCell-">setBackgroundColor(MenuItem, LabelCell)</a></span> - Method in class deltadak.<a href="../deltadak/Controller.html" title="class in deltadak">Controller</a></dt>
+<dt><span class="memberNameLink"><a href="../deltadak/Controller.html#setBackgroundColor-javafx.scene.control.MenuItem-deltadak.ui.LabelCell-">setBackgroundColor(MenuItem, LabelCell)</a></span> - Method in class deltadak.<a href="../deltadak/Controller.html" title="class in deltadak">Controller</a></dt>
 <dd>
 <div class="block">sets the background color of a LabelCell</div>
 </dd>
