+++ Javadoc/index-files/index-8.html
@@ -78,7 +78,7 @@
 <dd>
 <div class="block">custom ListCell</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../deltadak/LabelCell.html#LabelCell-deltadak.Controller-">LabelCell(Controller)</a></span> - Constructor for class deltadak.<a href="../deltadak/LabelCell.html" title="class in deltadak">LabelCell</a></dt>
+<dt><span class="memberNameLink"><a href="../deltadak/LabelCell.html#LabelCell-deltadak.ui.Controller-">LabelCell(Controller)</a></span> - Constructor for class deltadak.<a href="../deltadak/LabelCell.html" title="class in deltadak">LabelCell</a></dt>
 <dd>
 <div class="block">Constructor.</div>
 </dd>
