+++ Javadoc/index-files/index-9.html
@@ -82,7 +82,7 @@
 <dd>
 <div class="block">main method</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../deltadak/Controller.html#makeRepeatMenu-deltadak.LabelCell-java.time.LocalDate-">makeRepeatMenu(LabelCell, LocalDate)</a></span> - Method in class deltadak.<a href="../deltadak/Controller.html" title="class in deltadak">Controller</a></dt>
+<dt><span class="memberNameLink"><a href="../deltadak/Controller.html#makeRepeatMenu-deltadak.ui.LabelCell-java.time.LocalDate-">makeRepeatMenu(LabelCell, LocalDate)</a></span> - Method in class deltadak.<a href="../deltadak/Controller.html" title="class in deltadak">Controller</a></dt>
 <dd>
 <div class="block">Makes the menu with options to repeat for 1-8 weeks.</div>
 </dd>
