+++ src/deltadak/commands/DeleteCommand.java
@@ -1,5 +1,7 @@
-package deltadak;
+package deltadak.commands;
 
+import deltadak.ui.AbstractController;
+import deltadak.HomeworkTask;
 import javafx.scene.control.ListView;
 
 import java.time.LocalDate;
