+++ test/deltadak/ControllerJavaTest.java
@@ -1,5 +1,6 @@
 package deltadak;
 
+import deltadak.ui.Controller;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import org.junit.jupiter.api.Test;
